python/requirements.txt
opencv-python
numpy
numba
imgcat
Pillow